refactor2: split into files

refactor


refactor2


refactor3: add back calculate_ht_info and consider_probe_time_locality


s
