+++ be/src/exec/hash_join_components.cpp
@@ -23,7 +23,7 @@
 #include "common/logging.h"
 #include "common/object_pool.h"
 #include "exec/hash_joiner.h"
-#include "exec/join_hash_map.h"
+#include "exec/join/join_hash_map.h"
 #include "exprs/agg/distinct.h"
 #include "exprs/expr_context.h"
 #include "gutil/casts.h"
