+++ be/src/exec/hash_join_components.h
@@ -19,7 +19,7 @@
 
 #include "column/vectorized_fwd.h"
 #include "common/object_pool.h"
-#include "exec/join_hash_map.h"
+#include "exec/join/join_hash_map.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks {
