+++ be/src/exec/pipeline/hashjoin/spillable_hash_join_build_operator.cpp
@@ -23,7 +23,7 @@
 #include "exec/hash_join_components.h"
 #include "exec/hash_join_node.h"
 #include "exec/hash_joiner.h"
-#include "exec/join_hash_map.h"
+#include "exec/join/join_hash_map.h"
 #include "exec/pipeline/hashjoin/hash_join_build_operator.h"
 #include "exec/pipeline/hashjoin/hash_joiner_factory.h"
 #include "exec/pipeline/query_context.h"
