+++ be/test/exec/join_hash_map_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exec/join_hash_map.h"
+#include "exec/join/join_hash_map.h"
 
 #include <gtest/gtest.h>
